Changed from uint64_t to unsigned long long

+++ include/post.hpp
@@ -137,7 +137,7 @@ public:
 	/**
 	 * The post's unique ID.
 	 */
-	uint64_t id = 0;
+	unsigned long long id = 0;
 
 	/**
 	 * The post's unique ID as a String (For clients that don't support 64-bit integers).
@@ -167,7 +167,7 @@ public:
 	/**
 	 * The time of the post, in seconds since the epoch.
 	 */
-	uint64_t timestamp = 0;
+	unsigned long long timestamp = 0;
 
 	/**
 	 * The post format: html or markdown
@@ -217,7 +217,7 @@ public:
 	/**
 	 * The total number of post available for this request, useful for paginating through results.
 	 */
-	uint64_t total_posts = 0;
+	unsigned long long total_posts = 0;
 
 	/**
 	 * Short url for the post.
@@ -257,7 +257,7 @@ public:
 	/**
 	 * TODO Documentation
 	 */
-	uint64_t note_count = 0;
+	unsigned long long note_count = 0;
 
 	/**
 	 * TODO Documentation
